LayoutTests/http/tests/resourceLoadStatistics/omit-document-referrer-third-party-iframe-ephemeral-expected.txt

Tests that document.referrer is omitted in third-party iframes while in private browsing mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.referrer is [empty]
PASS successfullyParsed is true

TEST COMPLETE
